refactor: Product 컴포넌트를 도메인 타입으로 변경 #43
